refactor: split three-preview into modules
